"Remove unnecessary JSON file and methods, refactor student JSON
Deleted redundant 'students.json' file and removed 'JSONFileReader' class and 'saveJSON()' method as they are not required. Also, refactored the new 'students.json' file to match the data structure of our application. This is part of the effort to clean up our code base and ensure that we are only dealing with meaningful and necessary data."