"Replace pom.xml and add JSONWriter.java, update JSONReader.java
The pom.xml file was removed as we shifted from Maven builds to Gradle. A new file, JSONWriter.java, was added to handle writing of JSON files. Existing JSONReader.java was updated to remove print statements and to improve efficiency.

Changes made to enhance performance and readability of the code. Replacing Maven with Gradle for more finer-grained control on project configuration and dependencies. JSONWriter will help with creation/updation of JSON objects in a programmatic way. Comments in JSONReader debugging removed to maintain clean code standards."