Updated Registration.java file to include newline
This commit simply adds a newline at the end of the Registration.java file. This was done to make the file conform to best practice conventions which recommend files should end with a newline.

Having a newline at the end of each file allows utilities like diff to handle comparisons properly and it can also prevent certain issues with version control systems.